requirements.txt
requests==2.24.0
zeep==3.4.0
prometheus_client==0.8.0
behave==1.2.6
PyYAML==5.3.1
exchangelib~=3.2.0
allure-behave==2.8.18
selenium==3.141.0
cx_Oracle==8.2
clr==1.0.3
cryptography==3.1.1